Fixing code warnings, updating references, adjusting naming
