--- sampleCode/Python/endpoints/DataSetEndpoints.py
@@ -1,52 +1,0 @@
-# MIT License
-
-# Copyright (c) 2023 IMPLAN
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-import requests
-import json
-import logging
-
-class DataSet:
-    def __init__(self, id_, description, is_default):
-        self.id = id_
-        self.description = description
-        self.is_default = is_default
-
-class DataSetEndpoints:
-    @staticmethod
-    def get_datasets(aggregation_scheme_id, bearer_token):
-        url = f"https://api.implan.com/api/v1/datasets/{aggregation_scheme_id}"
-        headers = {"Authorization": f"Bearer {bearer_token}"}
-        
-        response = requests.get(url, headers=headers)
-        
-        if response.status_code == 200:
-            data = response.json()
-            logging.info(f"Data Sets: {data}")  # Log the data to see the exact keys
-            return [DataSet(**item) for item in data]
-        else:
-            logging.error(f"Failed to get data sets: {response.status_code} - {response.text}")
-            response.raise_for_status()
-
-# Setup basic logging configuration
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
-
-
