--- sampleCode/Python/endpoints/SpecificationEndpoints.py
@@ -1,42 +1,0 @@
-# MIT License
-
-# Copyright (c) 2023 IMPLAN
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-import requests
-
-class Specification:
-    def __init__(self, name, code):
-        self.name = name
-        self.code = code
-
-class SpecificationEndpoints:
-    @staticmethod
-    def get_specifications(project_id, event_type, bearer_token):
-        url = f"https://api.implan.com/api/v1/impact/project/{project_id}/eventtype/{event_type}/specification"
-        headers = {"Authorization": f"Bearer {bearer_token}"}
-        
-        response = requests.get(url, headers=headers)
-        if response.status_code == 200:
-            response_data = response.json()
-            specifications = [Specification(**item) for item in response_data]
-            return specifications
-        else:
-            response.raise_for_status()
