+++ sampleCode/Python/main.py
@@ -1,32 +1,11 @@
-# MIT License
-
-# Copyright (c) 2023 IMPLAN
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
 import logging
-from workflows.authentication_workflow import AuthenticationWorkflow
-from workflows.create_project_workflow import CreateProjectWorkflow
-from workflows.regional_workflow import RegionalWorkflow
-from workflows.combined_region_workflow import CombinedRegionWorkflow
-from workflows.MultiEventToMultiGroupWorkflow import MultiEventToMultiGroupWorkflow
-from workflows.run_impact_analysis_workflow import RunImpactAnalysisWorkflow
+
+from workflows.AuthenticationWorkflow import AuthenticationWorkflow
+from workflows.CreateProjectWorkflow import CreateProjectWorkflow
+from workflows.RegionalWorkflow import RegionalWorkflow
+# from workflows.CombinedRegionWorkflow import CombinedRegionWorkflow
+# from workflows.MultiEventToMultiGroupWorkflow import MultiEventToMultiGroupWorkflow
+from workflows.RunImpactAnalysisWorkflow import RunImpactAnalysisWorkflow
 
 def main():
     # Authenticate and get bearer token
