--- sampleCode/Python/services/desktop.ini
@@ -1,3 +1,0 @@
-[LocalizedFileNames]
-Implan_Logging.py=@Implan_Logging,0
-Implan_rest.py=@Implan_rest,0
