+++ sampleCode/Python/services/implan_logging.py
@@ -1,34 +1,12 @@
-# MIT License
-
-# Copyright (c) 2023 IMPLAN
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
 import os
 import sys
-import json
 import requests
 import datetime
+
 from pathlib import Path
-from typing import Optional, Dict, Any, Type
+from typing import Optional, Any, Type
 from requests.models import Request, Response
-from pydantic import json as pjson
+from pydantic import json
 
 class Logging:
     _log_file_path = None
@@ -75,7 +53,7 @@ class Logging:
 
         if request.data:
             log.append(f"-Body: '{request.headers.get('Content-Type')}'")
-            body_content = pjson.dumps(request.data, indent=2)
+            body_content = json.dumps(request.data, indent=2)
             log.append(body_content)
 
         # Response
@@ -93,7 +71,7 @@ class Logging:
                     log.append('')
                 
                 if response_data:
-                    response_json = pjson.dumps(response_data, indent=2)
+                    response_json = json.dumps(response_data, indent=2)
                     log.append(response_json)
                 else:
                     log.append(response.content.decode('utf-8'))
