+++ sampleCode/Python/services/implan_rest.py
@@ -1,33 +1,8 @@
-# MIT License
-
-# Copyright (c) 2023 IMPLAN
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
 import requests
-from requests.auth import AuthBase
-import json
 import logging
-from datetime import datetime
-from time import time
-from functools import wraps
-from pydantic import json as pjson
+
+from requests.auth import AuthBase
+from pydantic import json
 
 class BearerAuth(AuthBase):
     """Custom authentication class for bearer token."""
@@ -92,7 +67,7 @@ class RestClient:
     def get_response_data(response, data_type=None):
         try:
             return response.json()
-        except pjson.JSONDecodeError as e:
+        except json.JSONDecodeError as e:
             logging.error(f"JSON decode error: {e}")
             return None
