+++ sampleCode/Python/workflows/CombinedRegionWorkflow.py
@@ -1,31 +1,9 @@
-# MIT License
-
-# Copyright (c) 2023 IMPLAN
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
 import time
+
 from datetime import datetime
-from endpoints.AggregationSchemeEndpoints import AggregationSchemeEndpoints
-from endpoints.DataSetEndpoints import DataSetEndpoints
-from endpoints.Regions.RegionEndpoints import RegionEndpoints
-from workflows.authentication_workflow import AuthenticationWorkflow
+from endpoints.aggregation_scheme_endpoints import AggregationSchemeEndpoints
+from endpoints.dataset_endpoints import DataSetEndpoints
+from endpoints.Regions.region_endpoints import RegionEndpoints
 from workflows.iworkflow import IWorkflow
 from endpoints.Regions.CombineRegionRequest import CombineRegionRequest
