+++ sampleCode/Python/workflows/CreateProjectWorkflow.py
@@ -1,43 +1,15 @@
-# MIT License
-
-# Copyright (c) 2023 IMPLAN
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
 import logging
 import datetime
-from endpoints.AggregationSchemeEndpoints import AggregationSchemeEndpoints,AggregationScheme
-from endpoints.DataSetEndpoints import DataSetEndpoints,DataSet
-from endpoints.ProjectEndpoints import Project, ProjectEndpoints
-from endpoints.industry_sets import IndustrySet,IndustrySets
-from endpoints.IndustryCodeEndpoints import IndustryCodeEndpoints,IndustryCode
 
-from endpoints.Events.Event import Event
-from endpoints.Events.EventEndpoints import EventEndpoints
-# from endpoints.Events.IndustryOutputEvent import IndustryOutputEvent
+from endpoints.aggregation_scheme_endpoints import AggregationSchemeEndpoints
+from endpoints.dataset_endpoints import DataSetEndpoints
+from endpoints.project_endpoints import Project, ProjectEndpoints
+from endpoints.industry_set_endpoints import IndustrySets
+from endpoints.industry_code_endpoints import IndustryCodeEndpoints
+from endpoints.Events.event_endpoints import EventEndpoints
 from endpoints.Events import IndustryOutputEvent, IndustryImpactAnalysisEvent
-
-
-from endpoints.GroupEndpoints import GroupEndpoints,Group,GroupEvent
-from endpoints.Regions import CombineRegionRequest,Region
-from endpoints.Regions.RegionEndpoints import RegionEndpoints
-
+from endpoints.group_endpoints import GroupEndpoints,Group,GroupEvent
+from endpoints.Regions.region_endpoints import RegionEndpoints
 from workflows.iworkflow import IWorkflow
 
 class CreateProjectWorkflow(IWorkflow):
