--- sampleCode/Python/workflows/authentication_workflow.py
@@ -1,47 +1,0 @@
-# MIT License
-
-# Copyright (c) 2023 IMPLAN
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-# authentication_workflow.py
-from endpoints.authentication import ImplanAuthentication, Authentication
-import logging
-from workflows.iworkflow import IWorkflow
-
-class AuthenticationWorkflow(IWorkflow):
-    @staticmethod
-    def examples():
-        # Example Implan Authentication
-        auth = ImplanAuthentication(
-            username="",
-            password=""
-        )
-       
-        # Retrieve the token
-        bearer_token = Authentication.get_bearer_token(auth)
-
-        logging.info(f"Retrieved Bearer Token: {bearer_token}")
-
-        # Extract the actual token if it has "Bearer" prefix
-        if bearer_token.startswith("Bearer "):
-            bearer_token = bearer_token[len("Bearer "):]
-
-        # Return the bearer token for use in other workflows
-        return bearer_token
